refactor: extract startup ready lifecycle boundary
