+++ rustfs/src/lib.rs
@@ -71,6 +71,7 @@ pub mod server;
 pub mod startup_entrypoint;
 pub mod startup_fs_guard;
 pub mod startup_iam;
+pub mod startup_lifecycle;
 pub mod startup_optional_runtimes;
 pub mod startup_preflight;
 pub mod startup_profiling;
